# conecteai/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Conecte.ai",
        "version": "1.0"
    },
    "paths": {
        "/api/customer": {
            "get": {
                "tags": [
                    "CustomerController"
                ],
                "summary": "Lista os clientes cadastrados.",
                "description": "Lista os clientes cadastrados no banco de dados.",
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "content": {
                            "application/json": {
                                "example": [
                                    "$customer"
                                ]
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "CustomerController"
                ],
                "summary": "Atualiza um cliente.",
                "description": "Atualiza um novo cliente no banco de dados.",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Felipe Kazuo"
                    },
                    {
                        "name": "cnpj",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "99.999.999/9999-99"
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "03/03/2021"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "email@domain.com"
                    },
                    {
                        "name": "telephone",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "(99) 99999-9999"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "example": [
                                    "Erro ao atualizar cliente."
                                ]
                            }
                        }
                    },
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "example": [
                                    "Cliente atualizado com sucesso."
                                ]
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "CustomerController"
                ],
                "summary": "Cadastra um cliente.",
                "description": "Cadastra um novo cliente no banco de dados.",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Felipe Kazuo"
                    },
                    {
                        "name": "cnpj",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "99.999.999/9999-99"
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "03/03/2021"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "email@domain.com"
                    },
                    {
                        "name": "telephone",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "(99) 99999-9999"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "example": [
                                    "Erro ao cadastrar cliente."
                                ]
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "example": [
                                    "Erro ao cadastrar cliente."
                                ]
                            }
                        }
                    },
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "example": [
                                    "Cliente efetuado com sucesso."
                                ]
                            }
                        }
                    }
                }
            }
        },
        "/api/customer/{id}": {
            "get": {
                "tags": [
                    "CustomerController"
                ],
                "summary": "Mostra um cliente cadastrado.",
                "description": "Mostra um cliente cadastrado no banco de dados.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "example": [
                                    "Cliente não encontrado."
                                ]
                            }
                        }
                    },
                    "200": {
                        "description": "Successful Operation",
                        "content": {
                            "application/json": {
                                "example": [
                                    "$id"
                                ]
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "CustomerController"
                ],
                "summary": "Deleta um cliente.",
                "description": "Deleta um cliente existente no banco de dados.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "example": [
                                    "Erro ao deletar cliente."
                                ]
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "example": [
                                    "Cliente não encontrado."
                                ]
                            }
                        }
                    },
                    "200": {
                        "description": "Successful Operation",
                        "content": {
                            "application/json": {
                                "example": [
                                    "Cliente deletado com sucesso."
                                ]
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Faça login com email e senha para obter seu token.",
                "name": "Token",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}